+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add sub-commands project/task to edit command to open editor at line corresponding to project/task
 - Add edit flag to describe sub-commands to open up editor
 - Run tasks in parallell by default and add flag serial to opt out
+- Sync projects in parallell by default and add flag serial to opt out
 
 ## v0.6.1